new prodject structure 28.03.2025
