+++ src/main/java/wbd/api_client/ApiClient_GetAllOffers.java
